scanners/dns-scanner/requirements.txt
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
checkdmarc==4.4.4
dkimpy==1.0.5
dnspython==2.2.1
expiringdict==1.2.2
filelock==3.8.0
idna==3.4
nats-py==2.2.0
publicsuffix2==2.20191221
pycparser==2.21
pyleri==1.4.1
PyNaCl==1.5.0
python-dotenv==0.21.0
requests==2.28.1
requests-file==1.5.1
six==1.16.0
timeout-decorator==0.5.0
tldextract==3.4.0
urllib3==1.26.12
